correccionEnCompras
para plastico PEHD

+++ Sicapp/views.py
@@ -86,7 +86,7 @@ def compras(request):
             if cantidad!="0" and cantidad!= None:
                 concepto=request.POST.get("concepto"+str(i))
                 precio=request.POST.get("precio"+str(i))
-                subTotal = Decimal(precio)*int(float(cantidad))
+                subTotal = (float(precio)*int(float(cantidad)))
                 cantidadProd=cantidadProd+int(float(cantidad))
                 total=total+subTotal
                 detalle=Detallecompra(cantidad=cantidad,concepto=concepto,precio=precio,total=subTotal,compra=idCompra)
